fix(plugins/sigil): fix compilation on Windows
The local receiver depends on Unix process semantics (flock, Setsid,
signal-based liveness/termination, ps) that have no portable Windows
equivalent, so plugins/sigil failed to cross-compile for Windows.

Split the daemon internals behind build tags: daemon_unix.go keeps the
real implementation unchanged, daemon_windows.go stubs every entry point
to return a "not supported on Windows" sentinel.
